enum serialization, return stream safely
