+++ parea/wrapper/wrapper.py
@@ -185,7 +185,7 @@ class Wrapper:
 
     def _cleanup_trace_core(self, trace_id: str, start_time: datetime, error: str, cache_hit, args, kwargs):
         trace_data.get()[trace_id].cache_hit = cache_hit
-
+        parent_id = trace_context.get()[-2]
         if error:
             trace_data.get()[trace_id].error = error
             trace_data.get()[trace_id].status = "error"
@@ -197,7 +197,6 @@ class Wrapper:
             trace_data.get()[trace_id].end_timestamp = end_time.isoformat()
             trace_data.get()[trace_id].latency = (end_time - start_time).total_seconds()
 
-            parent_id = trace_context.get()[-2]
             trace_data.get()[parent_id].end_timestamp = end_time.isoformat()
             start_time_parent = datetime.fromisoformat(trace_data.get()[parent_id].start_timestamp)
             trace_data.get()[parent_id].latency = (end_time - start_time_parent).total_seconds()
@@ -207,7 +206,10 @@ class Wrapper:
 
             self.log(trace_id)
             self.log(parent_id)
-            trace_context.get().pop()
+            try:
+                trace_context.get().pop()
+            except IndexError:
+                pass
 
         return final_log
